+++ src/imio/email/parser/parser.py
@@ -116,6 +116,10 @@ class Parser:
             for part in payload:
                 if part.get_content_type() == "message/rfc822":  # maybe also check for attachment filename ?
                     self.origin = "Agent forward"
+                    if part.get("Content-Transfer-Encoding") == "base64":
+                        return email.message_from_bytes(
+                            base64.b64decode(part.get_payload()[0].as_string()), policy=email_policy
+                        )
                     return part.get_payload()[0]
             # ibm notes in base64
             if "IBM Notes" in message.get("X-Mailer", ""):